tests/wpt/tests/server-timing/navigation-timing-trickle.https.html

<!DOCTYPE html>
<head>
    <meta charset='utf-8' />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      setup({explicit_done: true})
      window.addEventListener("message", e => {
        test(() => {
          assert_equals(e.data, "PASS", "The iframe saw ServerTiming entries in its NavigationTiming entry.");
          done();
        })
      });
    </script>
</head>
<body>
<iframe src="resources/navigation-timing.html?pipe=trickle(800:d1)"></iframe>
</body>
