LayoutTests/imported/w3c/web-platform-tests/server-timing/navigation-timing-trickle.https-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'performance.getEntriesByType('navigation')[0].serverTiming')


Harness Error (TIMEOUT), message = null
